splitting out OutputOptions from CommonOptions
