+++ detect_secrets_server/actions/initialize.py
@@ -50,18 +50,20 @@ def initialize(args):
         for repo in args.repo
     ]
 
-    cron_repos = [repo for repo in tracked_repos if _clone_and_save_repo(repo)]
-    if not cron_repos:
-        return
+    for repo in tracked_repos:
+        _clone_and_save_repo(repo)
+    # cron_repos = [repo for repo in tracked_repos if _clone_and_save_repo(repo)]
+    # if not cron_repos:
+        # return
 
-    output = '# detect-secrets scanner'
-    for repo in cron_repos:
-        output += '\n{} {}'.format(
-            repo.cron(),
-            args.output_hook_command,
-        )
+    # output = '# detect-secrets scanner'
+    # for repo in cron_repos:
+        # output += '\n{} {}'.format(
+        # repo.cron(),
+        # args.output_hook_command,
+        # )
 
-    return output
+    # return output
 
 
 def _create_single_tracked_repo(
